Try fixing some errors

+++ dsp/src/Pico.c
@@ -7,6 +7,7 @@
 // For commercial use, separate licencing terms must be obtained.
 
 
+#include "teak/teak.h"
 #include "PicoInt.h"
 
 #ifdef SW_SCAN_RENDERER
@@ -15,19 +16,19 @@
 #define VIDEO_OPT 0x10
 #endif
 
-int PicoVer=0x0080;
+u16 PicoVer=0x0080;
 struct Pico Pico;
 // int PicoOpt=0xffffffff; // enable everything
 // int PicoOpt=0x10; // frame-based rendering
 // int PicoOpt = 0; // enable nothing, scanline rendering (broken in new ver)
 // int PicoOpt=0x17; // mono sound 
 // int PicoOpt=0x1f; // stereo sound
-int PicoOpt = VIDEO_OPT;
-int PicoSkipFrame=0; // skip rendering frame?
+u16 PicoOpt = VIDEO_OPT;
+u16 PicoSkipFrame=0; // skip rendering frame?
 
-static int DefaultCram(int cram)
+static u16 DefaultCram(u16 cram)
 {
-  int high=0x0841;
+  u16 high=0x0841;
   // Convert 0000bbbb ggggrrrr
   // to      rrrr1ggg g10bbbb1
   high|=(cram&0x00f)<<12; // Red
@@ -37,7 +38,7 @@ static int DefaultCram(int cram)
 }
 
 // Cram directly to BMP16 BG values for writing to an exrot BG
-static int BGCram(int cram)
+static u16 BGCram(u16 cram)
 {
 	// Convert	0000bbbbggggrrrr
 	// to		1bbbb0gggg0rrrr0
@@ -50,5 +51,5 @@ static int BGCram(int cram)
 }
 
 // Function to convert Megadrive Cram into a native colour:
-// int (*PicoCram)(int cram)=DefaultCram;
-int (*PicoCram)(int cram)=BGCram;
+// u16 (*PicoCram)(int cram)=DefaultCram;
+u16 (*PicoCram)(u16 cram)=BGCram;
